move file to src folder
